Datos mockeados para interfaz de usuario, admin, coach
