+++ src/data.js
@@ -1,0 +1,144 @@
+export const userBox = [
+    {
+      id: 1,
+      dni: '1117185534',
+      dniType: 'CC',
+      name: 'Luisa',
+      lastname: 'Tobón',
+      birthday: '12/04/2000',
+      email: 'luisatobon@email.com',
+      address: 'Calle 21 #30-72',
+      neighborhood: 'El Edén',
+      phone: '3214587623',
+      height: 160,
+      weight: 62,
+      profilePicture: 'LuisaTobonPhoto.png'
+    },
+    {
+      id: 2,
+      dni: '11234004',
+      dniType: 'CC',
+      name: 'Francisco',
+      lastname: 'Restrepo',
+      birthday: '11/07/1995',
+      email: 'franciscorestrepo@email.com',
+      address: 'Carrera 50 #32-70',
+      neighborhood: 'El Rosario',
+      phone: '3214596334',
+      height: 173,
+      weight: 70,
+      profilePicture: 'FranciscoRestrepoPhoto.png'
+    },
+    {
+      id: 3,
+      dni: '178200984',
+      dniType: 'CC',
+      name: 'Valeria',
+      lastname: 'Rendón',
+      birthday: '29/12/1989',
+      email: 'valeriarendon@email.com',
+      address: 'Calle 90 #31-72',
+      neighborhood: 'El Barco',
+      phone: '3214239878',
+      height: 160,
+      weight: 83,
+      profilePicture: 'ValeriaRendonPhoto.png'
+    },
+  ];
+  export const coach = [
+    {
+      id: 1,
+      dni: '1198452900',
+      dniType: 'CC',
+      name: 'Pablo',
+      lastname: 'Martínez',
+      birthday: '11/07/1980',
+      email: 'pablomartinez@email.com',
+      phone: '3214554333',
+      profilePicture: 'PabloMartinezPhoto.png'
+    },
+    {
+      id: 2,
+      dni: '90452900',
+      dniType: 'CC',
+      name: 'Daniela',
+      lastname: 'Blandón',
+      birthday: '26/02/1987',
+      email: 'danielablandon@email.com',
+      phone: '3006754456',
+      profilePicture: 'DanielaBlandonPhoto.png'
+    },
+  ];
+  export const user = [
+    {
+      id: 1,
+      dni: '1117185534',
+      password: 'password',
+      userTypeCode: '01'
+    },
+    {
+      id: 2,
+      dni: '11234004',
+      password: 'password',
+      userTypeCode: '01'
+    },
+    {
+      id: 3,
+      dni: '178200984',
+      password: 'password',
+      userTypeCode: '01'
+    },
+    {
+      id: 4,
+      dni: '1198452900',
+      password: 'password',
+      userTypeCode: '02'
+    },
+    {
+      id: 4,
+      dni: '90452900',
+      password: 'password',
+      userTypeCode: '02'
+    },
+    {
+      id: 5,
+      dni: '100100',
+      password: 'password',
+      userTypeCode: '03'
+    },
+  ];
+  export const userType = [
+    {
+      id: 1,
+      userTypeCode: '01',
+      userTypeName: 'Usuario'
+    },
+    {
+      id: 2,
+      userTypeCode: '02',
+      userTypeName: 'Coach'
+    },
+    {
+      id: 3,
+      userTypeCode: '03',
+      userTypeName: 'Administrador'
+    },
+  ];
+  export const plan = [
+    {
+      id: 1,
+      planCode: '8820',
+      planName: 'Tiquetera20',
+      validity: 40,
+      price: '100.000',
+      credits: 20,
+    },
+    {
+      id: 2,
+      planCode: '8830',
+      planName: 'Tiquetera30',
+      validity: 60,
+      price: '150.000',
+      credits: 30
+    },
+  ];
